pyproject.toml
[tool.poetry]
name = "pillowpractice-1"
version = "0.1.0"
description = ""
authors = ["godsmith99x <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
